adjusted license iinfo
